Undo moving tests
