+++ packages/synthetix-main/test/integration/modules/CollateralModuleConfiguration.test.ts
@@ -4,9 +4,9 @@ import assertRevert from '@synthetixio/core-utils/utils/assertions/assert-revert
 import { ethers as Ethers } from 'ethers';
 import { ethers } from 'hardhat';
 
-import { bootstrap } from '../../../bootstrap';
+import { bootstrap } from '../bootstrap';
 
-describe.only('CollateralManagerConfiguration (SCCP)', function () {
+describe('CollateralManagerConfiguration (SCCP)', function () {
   const { signers, systems } = bootstrap();
 
   let systemOwner: Ethers.Signer, user1: Ethers.Signer;
